refactor r / w hadoop methods
