+++ spark/src/main/scala/geotrellis/spark/io/hadoop/HadoopRasterMethods.scala
@@ -19,67 +19,10 @@ package geotrellis.spark.io.hadoop
 import geotrellis.util.MethodExtensions
 
 import org.apache.hadoop.conf.Configuration
-import org.apache.hadoop.fs.{FileSystem, Path}
-import org.apache.hadoop.io.compress.CompressionCodecFactory
+import org.apache.hadoop.fs.Path
 import org.apache.spark.SparkContext
-import java.io.{DataInputStream, DataOutputStream}
-
 
 trait HadoopRasterMethods[T] extends MethodExtensions[T] {
   def write(path: Path)(implicit sc: SparkContext): Unit = write(path, sc.hadoopConfiguration)
   def write(path: Path, conf: Configuration): Unit
-}
-
-object HadoopRasterMethods {
-  def write(path: Path, conf: Configuration)(dosWrite: DataOutputStream => Unit): Unit = {
-    val fs = FileSystem.get(conf)
-
-    val os = {
-      val factory = new CompressionCodecFactory(conf)
-      val codec = factory.getCodec(path)
-
-      if (codec == null) {
-        println(s"No codec found for $path, writing without compression.")
-        fs.create(path)
-      } else {
-        codec.createOutputStream(fs.create(path))
-      }
-    }
-    try {
-      val dos = new DataOutputStream(os)
-      try {
-        dosWrite(dos)
-      } finally {
-        dos.close
-      }
-    } finally {
-      os.close
-    }
-  }
-
-  def read[T](path: Path, conf: Configuration)(disRead: DataInputStream => T): T = {
-    val fs = FileSystem.get(conf)
-
-    val is = {
-      val factory = new CompressionCodecFactory(conf)
-      val codec = factory.getCodec(path)
-
-      if (codec == null) {
-        println(s"No codec found for $path, reading without compression.")
-        fs.open(path)
-      } else {
-        codec.createInputStream(fs.open(path))
-      }
-    }
-    try {
-      val dis = new DataInputStream(is)
-      try {
-        disRead(dis)
-      } finally {
-        dis.close
-      }
-    } finally {
-      is.close
-    }
-  }
 }
